Replace server-side image generation with client-side html-to-image to ensure reliable downloads
