+++ components/tarot/interpretation/action.tsx
@@ -53,6 +53,8 @@ interface ActionSectionProps {
 
 import { useTranslations } from "next-intl"
 
+import { toPng } from "html-to-image"
+
 export default function ActionSection({
     question: propQuestion,
     cards: propCards,
@@ -110,6 +112,25 @@ export default function ActionSection({
     useEffect(() => {
         void loadVersions()
     }, [loadVersions])
+
+    const imageRef = useRef<HTMLDivElement>(null)
+
+// Helper to get card image slug
+    function slugifyCardName(raw: string): { slug: string; isReversed: boolean } {
+        if (!raw) return { slug: "", isReversed: false }
+        const lower = raw.toLowerCase()
+        const isReversed =
+            lower.includes("(reversed)") || /\breversed\b/.test(lower)
+
+        const slug = lower
+            .replace(/\s*\(reversed\)/g, "")
+            .replace(/\s*reversed/g, "")
+            .trim()
+            .replace(/\s+/g, "-")
+            .replace(/[^a-z0-9-]/g, "")
+
+        return { slug, isReversed }
+    }
 
     const { complete } = useCompletion({
         api: "/api/interpret-cards/question",
@@ -563,6 +584,395 @@ export default function ActionSection({
 
     return (
         <div className='relative overflow-hidden group'>
+            {/* Hidden DOM element for image generation - positioned absolute off-screen */}
+            <div
+                style={{
+                    position: "absolute",
+                    top: -9999,
+                    left: -9999,
+                    width: 1080,
+                    height: 1350,
+                    overflow: "hidden",
+                }}
+            >
+                <div
+                    ref={imageRef}
+                    style={{
+                        width: "100%",
+                        height: "100%",
+                        display: "flex",
+                        flexDirection: "column",
+                        padding: 64,
+                        background:
+                            "radial-gradient(1400px 900px at 50% 0%, rgba(99,102,241,0.26) 0%, rgba(168,85,247,0.18) 35%, rgba(10,8,26,1) 72%), radial-gradient(1200px 900px at 50% 100%, rgba(234,179,8,0.18) 0%, rgba(10,8,26,1) 60%)",
+                        color: "#ffffff",
+                        fontFamily:
+                            "ui-sans-serif, system-ui, -apple-system, Segoe UI, Roboto, Ubuntu, Cantarell, Noto Sans, Helvetica Neue, Arial",
+                        position: "relative",
+                    }}
+                >
+                    {/* soft glow blobs */}
+                    <div
+                        style={{
+                            position: "absolute",
+                            top: -140,
+                            left: -160,
+                            width: 520,
+                            height: 520,
+                            borderRadius: 9999,
+                            background:
+                                "radial-gradient(circle at 30% 30%, rgba(234,179,8,0.40), rgba(234,179,8,0.00) 60%)",
+                            filter: "blur(24px)",
+                            opacity: 0.6,
+                        }}
+                    />
+                    <div
+                        style={{
+                            position: "absolute",
+                            bottom: -220,
+                            right: -200,
+                            width: 720,
+                            height: 720,
+                            borderRadius: 9999,
+                            background:
+                                "radial-gradient(circle at 30% 30%, rgba(56,189,248,0.32), rgba(56,189,248,0.00) 60%)",
+                            filter: "blur(30px)",
+                            opacity: 0.55,
+                        }}
+                    />
+
+                    {/* background card aura */}
+                    {(cards || [])
+                        .slice(0, 3)
+                        .map((cName) => {
+                            const { slug, isReversed } = slugifyCardName(cName)
+                            // Use direct path for client side
+                            const src = `/assets/rider-waite-tarot/${slug}.png`
+                            return { name: cName, slug, isReversed, src }
+                        })
+                        .map((c, idx) => {
+                            const positions: Array<{
+                                top?: number
+                                bottom?: number
+                                left?: number
+                                right?: number
+                                rotate: number
+                            }> = [
+                                { top: 120, left: 60, rotate: -14 },
+                                { top: 150, right: 80, rotate: 16 },
+                                { bottom: 560, left: 80, rotate: -10 },
+                            ]
+                            const p = positions[idx % positions.length]
+                            return (
+                                <img
+                                    key={`bg-${c.slug}-${idx}`}
+                                    src={c.src}
+                                    width={260}
+                                    height={420}
+                                    style={{
+                                        position: "absolute",
+                                        ...(p.top != null
+                                            ? { top: p.top }
+                                            : {}),
+                                        ...(p.bottom != null
+                                            ? { bottom: p.bottom }
+                                            : {}),
+                                        ...(p.left != null
+                                            ? { left: p.left }
+                                            : {}),
+                                        ...(p.right != null
+                                            ? { right: p.right }
+                                            : {}),
+                                        transform: `rotate(${p.rotate}deg) scale(0.9)`,
+                                        opacity: 0.14,
+                                    }}
+                                />
+                            )
+                        })}
+
+                    <div
+                        style={{
+                            display: "flex",
+                            flexDirection: "column",
+                            position: "relative",
+                            zIndex: 10,
+                        }}
+                    >
+                        {/* Brand */}
+                        <div
+                            style={{
+                                display: "flex",
+                                alignItems: "center",
+                                gap: 12,
+                                marginBottom: 26,
+                            }}
+                        >
+                            <div
+                                style={{
+                                    width: 44,
+                                    height: 44,
+                                    borderRadius: 9999,
+                                    background:
+                                        "linear-gradient(135deg, rgba(234,179,8,0.95), rgba(56,189,248,0.75))",
+                                    display: "flex",
+                                    alignItems: "center",
+                                    justifyContent: "center",
+                                    fontWeight: 900,
+                                    color: "#0a081a",
+                                    boxShadow:
+                                        "0 14px 40px rgba(234,179,8,0.22)",
+                                }}
+                            >
+                                <svg
+                                    width='24'
+                                    height='24'
+                                    viewBox='0 0 24 24'
+                                    fill='currentColor'
+                                    xmlns='http://www.w3.org/2000/svg'
+                                >
+                                    <path d='M12 2L15 9L22 12L15 15L12 22L9 15L2 12L9 9Z' />
+                                </svg>
+                            </div>
+                            <div
+                                style={{
+                                    fontSize: 30,
+                                    fontWeight: 900,
+                                    letterSpacing: -0.4,
+                                }}
+                            >
+                                Asking Fate
+                            </div>
+                        </div>
+
+                        {/* Question card */}
+                        <div
+                            style={{
+                                borderRadius: 28,
+                                padding: 34,
+                                border: "1px solid rgba(255,255,255,0.12)",
+                                background:
+                                    "linear-gradient(135deg, rgba(255,255,255,0.08), rgba(255,255,255,0.03) 45%, rgba(56,189,248,0.06) 90%)",
+                                boxShadow:
+                                    "0 18px 70px -30px rgba(56,189,248,0.55)",
+                            }}
+                        >
+                            <div
+                                style={{
+                                    fontSize: 20,
+                                    opacity: 0.85,
+                                    marginBottom: 12,
+                                }}
+                            >
+                                Your question
+                            </div>
+                            <div
+                                style={{
+                                    fontFamily:
+                                        "ui-serif, Georgia, Cambria, Times New Roman, Times, serif",
+                                    fontSize: 44,
+                                    fontWeight: 900,
+                                    lineHeight: 1.15,
+                                    textShadow:
+                                        "0 10px 30px rgba(56,189,248,0.22)",
+                                }}
+                            >
+                                {`“${
+                                    (question || "").length > 140
+                                        ? (question || "")
+                                              .slice(0, 139)
+                                              .trimEnd() + "…"
+                                        : question || ""
+                                }”`}
+                            </div>
+
+                            {/* Selected cards row */}
+                            <div
+                                style={{
+                                    display: "flex",
+                                    gap: 18,
+                                    flexWrap: "wrap",
+                                    marginTop: 26,
+                                    alignItems: "flex-start",
+                                }}
+                            >
+                                {(cards || [])
+                                    .slice(0, 3)
+                                    .map((cName) => {
+                                        const { slug, isReversed } =
+                                            slugifyCardName(cName)
+                                        const src = `/assets/rider-waite-tarot/${slug}.png`
+                                        return {
+                                            name: cName,
+                                            slug,
+                                            isReversed,
+                                            src,
+                                        }
+                                    })
+                                    .map((c, idx) => (
+                                        <div
+                                            key={`card-${c.slug}-${idx}`}
+                                            style={{
+                                                width: 170,
+                                                display: "flex",
+                                                flexDirection: "column",
+                                                gap: 10,
+                                                alignItems: "center",
+                                            }}
+                                        >
+                                            <div
+                                                style={{
+                                                    fontSize: 14,
+                                                    padding: "8px 10px",
+                                                    borderRadius: 9999,
+                                                    background:
+                                                        "rgba(255,255,255,0.12)",
+                                                    border: "1px solid rgba(99,102,241,0.22)",
+                                                    color: "rgba(255,255,255,0.92)",
+                                                    textAlign: "center",
+                                                    maxWidth: 170,
+                                                    overflow: "hidden",
+                                                    whiteSpace: "nowrap",
+                                                    textOverflow: "ellipsis",
+                                                }}
+                                            >
+                                                {c.name}
+                                            </div>
+
+                                            <div
+                                                style={{
+                                                    width: 150,
+                                                    height: 240,
+                                                    borderRadius: 18,
+                                                    position: "relative",
+                                                    overflow: "hidden",
+                                                    boxShadow:
+                                                        "0 20px 60px -35px rgba(234,179,8,0.65)",
+                                                    border: "1px solid rgba(255,255,255,0.12)",
+                                                }}
+                                            >
+                                                <div
+                                                    style={{
+                                                        position: "absolute",
+                                                        inset: -30,
+                                                        background:
+                                                            "radial-gradient(circle at 30% 20%, rgba(99,102,241,0.35), rgba(99,102,241,0.0) 55%), radial-gradient(circle at 70% 80%, rgba(234,179,8,0.25), rgba(234,179,8,0.0) 60%)",
+                                                        filter: "blur(18px)",
+                                                        opacity: 0.9,
+                                                    }}
+                                                />
+                                                <img
+                                                    src={c.src}
+                                                    width={150}
+                                                    height={240}
+                                                    style={{
+                                                        position: "absolute",
+                                                        inset: 0,
+                                                        objectFit: "cover",
+                                                        transform: c.isReversed
+                                                            ? "rotate(180deg)"
+                                                            : "rotate(0deg)",
+                                                    }}
+                                                />
+                                            </div>
+                                        </div>
+                                    ))}
+                            </div>
+                        </div>
+
+                        {/* Interpretation card */}
+                        <div
+                            style={{
+                                marginTop: 28,
+                                borderRadius: 28,
+                                padding: 32,
+                                background:
+                                    "linear-gradient(135deg, rgba(99,102,241,0.18), rgba(168,85,247,0.14) 35%, rgba(34,211,238,0.12) 70%)",
+                                boxShadow:
+                                    "0 20px 70px -35px rgba(56,189,248,0.55)",
+                                border: "1px solid rgba(255,255,255,0.12)",
+                                position: "relative",
+                            }}
+                        >
+                            <div
+                                style={{
+                                    display: "flex",
+                                    alignItems: "center",
+                                    gap: 12,
+                                    marginBottom: 14,
+                                }}
+                            >
+                                <div
+                                    style={{
+                                        width: 42,
+                                        height: 42,
+                                        borderRadius: 9999,
+                                        background: "rgba(234,179,8,0.18)",
+                                        border: "1px solid rgba(234,179,8,0.25)",
+                                        display: "flex",
+                                        alignItems: "center",
+                                        justifyContent: "center",
+                                        color: "rgba(255,255,255,0.9)",
+                                        fontWeight: 900,
+                                    }}
+                                >
+                                    <svg
+                                        width='24'
+                                        height='24'
+                                        viewBox='0 0 24 24'
+                                        fill='currentColor'
+                                        xmlns='http://www.w3.org/2000/svg'
+                                    >
+                                        <path d='M12 2L14.5 9.5L22 12L14.5 14.5L12 22L9.5 14.5L2 12L9.5 9.5Z' />
+                                    </svg>
+                                </div>
+                                <div
+                                    style={{
+                                        fontSize: 26,
+                                        fontWeight: 900,
+                                        letterSpacing: -0.2,
+                                    }}
+                                >
+                                    Your reading
+                                </div>
+                            </div>
+
+                            <div
+                                style={{
+                                    display: "block",
+                                    fontSize: 28,
+                                    lineHeight: 1.5,
+                                    whiteSpace: "pre-wrap",
+                                    color: "rgba(255,255,255,0.92)",
+                                }}
+                            >
+                                {`${
+                                    (interpretation || "").length > 900
+                                        ? (interpretation || "")
+                                              .slice(0, 899)
+                                              .trimEnd() + "…"
+                                        : interpretation || "—"
+                                }`}
+                            </div>
+                        </div>
+
+                        {/* Footer */}
+                        <div
+                            style={{
+                                display: "flex",
+                                justifyContent: "space-between",
+                                alignItems: "center",
+                                marginTop: 28,
+                                opacity: 0.85,
+                                fontSize: 18,
+                            }}
+                        >
+                            <div>Generated with Asking Fate</div>
+                            <div>askingfate.com</div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+
             {/* Slide-up loader for download */}
             {isDownloading && (
                 <div className='fixed bottom-0 left-0 right-0 z-50 animate-slide-up bg-black/60 backdrop-blur-sm p-4 text-center text-white'>
@@ -654,99 +1064,15 @@ export default function ActionSection({
                                                     type='button'
                                                     className='w-full px-3 py-2 rounded-md bg-primary/20 hover:bg-primary/30 text-sm'
                                                     onClick={async () => {
+                                                        if (!imageRef.current) return
                                                         try {
-                                                            setIsDownloading(
-                                                                true
-                                                            )
-                                                            const type = "image"
-                                                            const res =
-                                                                await fetch(
-                                                                    "/api/share-image",
+                                                            setIsDownloading(true)
+                                                            const dataUrl =
+                                                                await toPng(
+                                                                    imageRef.current,
                                                                     {
-                                                                        method: "POST",
-                                                                        headers:
-                                                                            {
-                                                                                "Content-Type":
-                                                                                    "application/json",
-                                                                            },
-                                                                        body: JSON.stringify(
-                                                                            {
-                                                                                question,
-                                                                                cards: cards,
-                                                                                interpretation,
-                                                                                // Safer "story" size (avoids huge/corrupted renders on some devices)
-                                                                                width: 1080,
-                                                                                height: 1920,
-                                                                                branding:
-                                                                                    "Asking Fate",
-                                                                                theme: "cosmic",
-                                                                                type,
-                                                                            }
-                                                                        ),
-                                                                    }
-                                                                )
-                                                            if (!res.ok) {
-                                                                const msg =
-                                                                    (await res
-                                                                        .text()
-                                                                        .catch(
-                                                                            () =>
-                                                                                ""
-                                                                        )) ||
-                                                                    "Failed to generate image"
-                                                                toast.error(
-                                                                    msg.slice(
-                                                                        0,
-                                                                        180
-                                                                    )
-                                                                )
-                                                                return
-                                                            }
-                                                            const ct =
-                                                                res.headers.get(
-                                                                    "content-type"
-                                                                ) || ""
-                                                            if (
-                                                                !ct.includes(
-                                                                    "image/png"
-                                                                )
-                                                            ) {
-                                                                toast.error(
-                                                                    "Share image returned an unsupported format. Please try again."
-                                                                )
-                                                                return
-                                                            }
-                                                            const buf =
-                                                                await res.arrayBuffer()
-                                                            const bytes = new Uint8Array(
-                                                                buf.slice(0, 8)
-                                                            )
-                                                            const pngMagic = [
-                                                                0x89, 0x50,
-                                                                0x4e, 0x47,
-                                                                0x0d, 0x0a,
-                                                                0x1a, 0x0a,
-                                                            ]
-                                                            const isPng =
-                                                                bytes.length ===
-                                                                    8 &&
-                                                                pngMagic.every(
-                                                                    (b, i) =>
-                                                                        bytes[
-                                                                            i
-                                                                        ] === b
-                                                                )
-                                                            if (!isPng) {
-                                                                toast.error(
-                                                                    "Downloaded file is not a valid PNG. Please try again."
-                                                                )
-                                                                return
-                                                            }
-                                                            const blob =
-                                                                new Blob(
-                                                                    [buf],
-                                                                    {
-                                                                        type: "image/png",
+                                                                        cacheBust: true,
+                                                                        pixelRatio: 2,
                                                                     }
                                                                 )
                                                             const ts =
@@ -756,28 +1082,27 @@ export default function ActionSection({
                                                                         /[:.]/g,
                                                                         "-"
                                                                     )
-                                                            const url =
-                                                                URL.createObjectURL(
-                                                                    blob
-                                                                )
                                                             const a =
                                                                 document.createElement(
                                                                     "a"
                                                                 )
-                                                            a.href = url
+                                                            a.href = dataUrl
                                                             a.download = `reading-${ts}.png`
                                                             document.body.appendChild(
                                                                 a
                                                             )
                                                             a.click()
                                                             a.remove()
-                                                            URL.revokeObjectURL(
-                                                                url
+                                                        } catch (err) {
+                                                            console.error(
+                                                                "Failed to generate image client-side:",
+                                                                err
+                                                            )
+                                                            toast.error(
+                                                                "Failed to generate image. Please try again."
                                                             )
                                                         } finally {
-                                                            setIsDownloading(
-                                                                false
-                                                            )
+                                                            setIsDownloading(false)
                                                         }
                                                     }}
                                                 >
